requirements.txt
# Kafka
kafka-python==2.0.2
confluent-kafka==2.8.2
websocket-client==1.6.4

# Spark
pyspark==3.5.0
findspark==2.0.1